+++ src/main/java/com/jovan/erp_v1/request/DeliveryItemOutboundRequest.java
@@ -1,5 +1,7 @@
 package com.jovan.erp_v1.request;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -7,6 +9,6 @@ import jakarta.validation.constraints.Positive;
 public record DeliveryItemOutboundRequest(
         Long id,
         @NotEmpty Long productId,
-        @NotNull @Positive Double quantity,
+        @NotNull @Positive BigDecimal quantity,
         @NotNull Long outboundDeliveryId) {
 }
