+++ src/main/java/com/jovan/erp_v1/request/DeliveryItemRequest.java
@@ -1,5 +1,7 @@
 package com.jovan.erp_v1.request;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -7,7 +9,7 @@ import jakarta.validation.constraints.Positive;
 public record DeliveryItemRequest(
         Long id,
         @NotEmpty Long productId,
-        @NotNull @Positive Double quantity,
+        @NotNull @Positive BigDecimal quantity,
         @NotNull Long inboundDeliveryId,
         @NotNull Long outboundDeliveryId) {
 }
